moved out components from hooks, moved out hooks
