+++ view/src/hooks/useCountUp.ts
@@ -1,0 +1,32 @@
+import * as React from 'react';
+export function useCountUp(target: number, duration: number = 1200, delay: number = 300) {
+  const [count, setCount] = React.useState(0);
+
+  React.useEffect(() => {
+    let startTime: number | null = null;
+    let rafId: number;
+
+    const timeout = setTimeout(() => {
+      const step = (timestamp: number) => {
+        if (!startTime) startTime = timestamp;
+        const elapsed = timestamp - startTime;
+        const progress = Math.min(elapsed / duration, 1);
+        const eased = 1 - Math.pow(1 - progress, 3);
+        setCount(Math.floor(eased * target));
+        if (progress < 1) {
+          rafId = requestAnimationFrame(step);
+        } else {
+          setCount(target);
+        }
+      };
+      rafId = requestAnimationFrame(step);
+    }, delay);
+
+    return () => {
+      clearTimeout(timeout);
+      cancelAnimationFrame(rafId);
+    };
+  }, [target, duration, delay]);
+
+  return count;
+}
